Move Dispatch and Execute API endpoints to Workflows resource
